@@ -14,21 +14,21 @@
     "prebuild": "rimraf ./dist && mkdirp ./dist",
     "postbuild": "npm run deploy",
     "predeploy": "rimraf ../../../backend/src/main/resources/static/*",
-    "deploy": "cpy 'base/*' dist/testproject/ && cpy 'dist/testproject/**' ../../../backend/src/main/resources/static/",
+    "deploy": "cpy 'base/*' dist/testproject/browser/ && cpy 'dist/testproject/browser/**' ../../../backend/src/main/resources/static/",
     "prettier": "npx prettier --write 'src/**/*.{ts,html,scss,json}'",
     "extract-i18n": "ng extract-i18n --format json"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.0",
-    "@angular/common": "^17.0.0",
-    "@angular/compiler": "^17.0.0",
-    "@angular/core": "^17.0.0",
-    "@angular/forms": "^17.0.0",
-    "@angular/localize": "^17.0.0",
-    "@angular/platform-browser": "^17.0.0",
-    "@angular/platform-browser-dynamic": "^17.0.0",
-    "@angular/router": "^17.0.0",
+    "@angular/animations": "^18.0.0",
+    "@angular/common": "^18.0.0",
+    "@angular/compiler": "^18.0.0",
+    "@angular/core": "^18.0.0",
+    "@angular/forms": "^18.0.0",
+    "@angular/localize": "^18.0.0",
+    "@angular/platform-browser": "^18.0.0",
+    "@angular/platform-browser-dynamic": "^18.0.0",
+    "@angular/router": "^18.0.0",
     "@ng-bootstrap/ng-bootstrap": "^16.0.0",
     "ngx-simple-charts": "^17.0.0",
     "@popperjs/core": "^2.11.8",
@@ -38,10 +38,10 @@
     "zone.js": "~0.14.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.1",
-    "@angular/cli": "^17.0.0",
-    "@angular/compiler-cli": "^17.0.0",
-    "@angular/language-service": "^17.0.0",
+    "@angular-devkit/build-angular": "^18.0.1",
+    "@angular/cli": "^18.0.1",
+    "@angular/compiler-cli": "^18.0.0",
+    "@angular/language-service": "^18.0.0",
     "@types/jasmine": "~3.10.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^14.0.0",
@@ -58,7 +58,7 @@
     "npm-run-all": "^4.1.5",
     "ts-node": "~10.4.0",
     "rimraf": "^3.0.2",
-    "typescript": "~5.2.0",
+    "typescript": "~5.4.5",
     "prettier": "^2.8.1",
     "webpack-bundle-analyzer": "^4.5.0",
     "eslint": "^8.0.0",
